// frontend/src/api-generated/models/danhSachThietBiPhanCungRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DanhSachThietBiPhanCungRequest {
  idTaiSanPhanCung: number;
  idNhaCungCap?: number;
  /**
     * @minLength 0
     * @maxLength 100
     */
  soSerial: string;
  /**
     * @minLength 0
     * @maxLength 50
     */
  maTheTaiSan?: string;
  /** @minimum 0 */
  giaMua?: number;
  thoiGianMua?: string;
  /** @minimum 0 */
  hanBaoHanhThang?: number;
  /**
     * @minLength 0
     * @maxLength 50
     */
  trangThaiKho?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  viTriKho?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}